[More testing...] Various fixes.
* Added flush to EventHandler (??)
* PojoOptions.getExperimenter() -> Long
* iQuery/iUpdate got various improvements (initialize())
* Null-pointer in CountCollector
* Filters are now decidedly NOT Thread-safe.
  (Now need a wrapper to use filters as interceptors)
* Lots of changes to /query !



git-svn-id: file:///home/svn/omero/trunk@437 05709c45-44f0-0310-885b-81a1db45b4a6

+++ components/common/src/ome/util/builders/PojoOptions.java
@@ -176,7 +176,7 @@ public class PojoOptions
      * Filtered by Experimenter
      * ============================== */
     
-    public PojoOptions exp(Integer i){
+    public PojoOptions exp(Long i){
     	options.put(EXPERIMENTER,i);
     	return this;
     }
@@ -190,8 +190,8 @@ public class PojoOptions
     	return options.containsKey(EXPERIMENTER);
     }
     
-    public Integer getExperimenter(){
-    	return (Integer) options.get(EXPERIMENTER);
+    public Long getExperimenter(){
+    	return (Long) options.get(EXPERIMENTER);
     }
 
     /* ==============================
@@ -199,7 +199,7 @@ public class PojoOptions
      * ============================== */
 
     /* FIXME : This is currently disabled. 
-    public PojoOptions grp(Integer i){
+    public PojoOptions grp(Long i){
         options.put(GROUP,i);
         return this;
     }
@@ -213,8 +213,8 @@ public class PojoOptions
         return options.containsKey(GROUP);
     }
     
-    public Integer getGroup(){
-        return (Integer) options.get(GROUP);
+    public Long getGroup(){
+        return (Long) options.get(GROUP);
     }
     */
